# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "O-Ushers API",
    "description": "The **O-Ushers** platform connects event organizers with professional ushers (talents).\n\n## Roles\n| Role | Description |\n|------|-------------|\n| **Usher** | A talent who browses events, applies, and manages their profile |\n| **Organizer** | Posts events, manages applicants, marks attendance & reviews |\n| **Organizer Member** | Uses the organizer workspace without event or staff CRUD access |\n| **Organizer Supervisor** | Uses the organizer workspace and manages event operations |\n| **Admin** | Platform-level management: users, events, verification |\n\n## Authentication\nBrowser clients authenticate with secure HTTP-only cookies. API clients may also send the\nshort-lived access token returned by `POST /auth/login` as a Bearer token.\n\n## Pagination\nPaginated endpoints accept `page` and `size` query parameters and return:\n```json\n{ \"data\": [], \"currentPage\": 1, \"totalPages\": 5, \"totalCount\": 42 }\n```\n",
    "version": "1.0.0",
    "contact": {
      "email": "ooushers@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://o-ushers.vercel.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/auth/forget-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request password reset OTP",
        "description": "Sends a 6-digit OTP to the user's email. OTP is valid for **15 minutes**.\nRequests are rate-limited to one per 30 seconds.\n",
        "security": [],
        "operationId": "forgetPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "ahmed@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent to email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "description": "Rate limit — wait before requesting a new OTP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify OTP",
        "description": "Validates the 6-digit OTP. After 3 failed attempts the OTP is\ninvalidated and a new one must be requested.\n",
        "security": [],
        "operationId": "verifyOtp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "otp"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "ahmed@example.com"
                  },
                  "otp": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 6,
                    "example": "483920"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified — proceed to reset-password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Max OTP attempts exceeded"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password",
        "description": "Sets a new password. Requires OTP to have been verified first.",
        "security": [],
        "operationId": "resetPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "newPassword"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "ahmed@example.com"
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "example": "NewStrongPass@456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "403": {
            "description": "OTP not verified"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get own profile (any role)",
        "description": "Returns the authenticated user's full profile. Works for all roles.",
        "operationId": "getMyProfile",
        "responses": {
          "200": {
            "description": "Own profile",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived JWT obtained from `POST /auth/login`. Browser clients normally\nrely on the secure HTTP-only session cookies instead.\n"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Validation error or bad input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Validation error: eventId is required"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid Bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "User not found"
            }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Something went wrong"
          }
        }
      },
...
      "CloudinaryFile": {
        "type": "object",
        "description": "Image metadata returned by Cloudinary after upload.",
        "properties": {
          "secure_url": {
            "type": "string",
            "format": "uri",
            "example": "https://res.cloudinary.com/example/image/upload/v123/ushers/profiles/abc.jpg"
          },
          "public_id": {
            "type": "string",
            "example": "ushers/profiles/abc"
          }
        }
      },
      "User": {
        "type": "object",
        "description": "User model. Sensitive fields (password, otp, otpExpiry, otpAttempts,\nlastOtpRequest, otpVerified) are never returned in responses.\n",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "6ba45f12-1234-4562-c3ab-9f963f44bfa4"
          },
          "fullName": {
            "type": "string",
            "example": "Ahmed Ali"
          },
          "userName": {
            "type": "string",
            "example": "ahmed_ali"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "ahmed@example.com"
          },
          "mobileNumber": {
            "type": "string",
            "example": "+201234567890"
          },
          "city": {
            "type": "string",
            "example": "Cairo"
          },
          "role": {
            "type": "string",
            "enum": [
              "usher",
              "organizer",
              "admin"
            ]
          },
          "experience": {
            "type": "integer",
            "example": 3
          },
          "rate": {
            "type": "number",
            "format": "float",
            "example": 4.7
          },
          "isVerified": {
            "type": "boolean",
            "example": true
          },
          "isBlocked": {
            "type": "boolean",
            "example": false
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "verified",
              "blocked"
            ],
            "example": "verified"
          },
          "lateExcuseCount": {
            "type": "integer",
            "example": 1
          },
          "consecutiveGoodEvents": {
            "type": "integer",
            "example": 3
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "arabic",
                "english",
                "french",
                "spanish",
                "german",
                "italian"
              ]
            },
            "example": [
              "arabic",
              "english"
            ]
          },
          "eventCategories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "wedding",
                "corporate",
                "club",
                "festival",
                "private_party",
                "conference",
                "exhibition",
                "sport_event"
              ]
            },
            "example": [
              "wedding",
              "corporate"
            ]
          },
          "portfolioPicture": {
            "$ref": "#/components/schemas/CloudinaryFile"
          },
          "portfolio": {
            "type": "array",
            "items": {}
          },
          "organizationInfo": {
            "type": "object",
            "nullable": true,
            "description": "Organizer-specific data (description, website).",
            "properties": {
              "description": {
                "type": "string"
              },
              "website": {
                "type": "string",
                "format": "uri"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}